Initial commit for python interface
The python interface has been designed as a 'calculator'
of the Atomistic Simulation Environment.
For info on how to install the interface, see <root>/python/INSTALL.

This implementation is a first draft.
It writes an input file in the original format for the ljmd-*.x binary
and then calls the binary on this input file.

For later versions, we may consider to change the file format,
since the line-dependent syntax is very unflexible.
The code for the interface will, of course, have to be adapted
accordingly.