Renamed MyMD class to FileIOMyMD
The current python interface can just read and write input files.
This was easy to start with, since it does not rely on the internal
structure of the C++ code.
However, it is less powerful than directly exposing the internal
structure to python via the cython module and this is the intended next
step.
The calculator class interfacing to cython will be called MyMD and
derive from ase.calculators.Calculator, while the calculator class that
reads input/output files is called FileIOMyMD and derives from
ase.calculators.FileIOCalculator